chat_client/requirements.txt
protobuf
grpcio
python-etcd
